# ReservationSystem/docs/api-specs/booking-openapi.yaml
openapi: 3.0.0
info:
  title: Booking Service API
  description: Hotel booking management API
  version: 1.0.0
  contact:
    name: API Support
    email: support@hotel.com

servers:
  - url: http://localhost:8083
    description: Local development
  - url: https://api.hotel.com/bookings
    description: Production

paths:
  /health:
    get:
      summary: Health check
      tags:
        - Health
      responses:
        '200':
          description: Service is healthy
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: healthy

  /bookings:
    get:
      summary: Get user bookings
      tags:
        - Bookings
      security:
        - BearerAuth: []
      parameters:
        - in: header
          name: user-id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: List of bookings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Booking'
        '401':
          description: Unauthorized

    post:
      summary: Create a new booking
      tags:
        - Bookings
      security:
        - BearerAuth: []
      parameters:
        - in: header
          name: user-id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BookingCreate'
      responses:
        '201':
          description: Booking created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Booking'
        '400':
          description: Invalid request
        '401':
          description: Unauthorized

  /bookings/{id}:
    get:
      summary: Get booking by ID
      tags:
        - Bookings
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
        - in: header
          name: user-id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Booking details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Booking'
        '404':
          description: Booking not found
        '403':
          description: Not authorized

    put:
      summary: Update booking
      tags:
        - Bookings
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
        - in: header
          name: user-id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BookingCreate'
      responses:
        '200':
          description: Booking updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Booking'
        '404':
          description: Booking not found

    delete:
      summary: Cancel booking
      tags:
        - Bookings
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
        - in: header
          name: user-id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Booking cancelled
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Booking cancelled successfully
        '404':
          description: Booking not found

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    BookingCreate:
      type: object
      required:
        - hotel_id
        - room_id
        - check_in
        - check_out
        - guests
        - total_price
      properties:
        hotel_id:
          type: integer
          example: 1
        room_id:
          type: integer
          example: 101
        check_in:
          type: string
          format: date-time
          example: "2024-01-15T14:00:00Z"
        check_out:
          type: string
          format: date-time
          example: "2024-01-20T11:00:00Z"
        guests:
          type: integer
          minimum: 1
          example: 2
        total_price:
          type: number
          format: float
          example: 1250.00

    Booking:
      type: object
      properties:
        id:
          type: integer
          example: 1
        user_id:
          type: integer
          example: 123
        hotel_id:
          type: integer
          example: 1
        room_id:
          type: integer
          example: 101
        check_in:
          type: string
          format: date-time
          example: "2024-01-15T14:00:00Z"
        check_out:
          type: string
          format: date-time
          example: "2024-01-20T11:00:00Z"
        guests:
          type: integer
          example: 2
        total_price:
          type: number
          format: float
          example: 1250.00
        status:
          type: string
          enum: [PENDING, CONFIRMED, CANCELLED, COMPLETED]
          example: CONFIRMED
        created_at:
          type: string
          format: date-time
          example: "2024-01-10T10:30:00Z"